+++ src/frontend/src/optimizer/plan_node/stream_simple_agg.rs
