+++ src/frontend/src/optimizer/plan_node/stream_stateless_simple_agg.rs
