+++ src/stream/src/executor/mod.rs
@@ -66,11 +66,9 @@ mod dynamic_filter;
 mod error;
 mod expand;
 mod filter;
-mod global_simple_agg;
 mod hash_agg;
 pub mod hash_join;
 mod hop_window;
-mod local_simple_agg;
 mod lookup;
 mod lookup_union;
 mod managed_state;
@@ -84,12 +82,14 @@ mod project_set;
 mod rearranged_chain;
 mod receiver;
 pub mod row_id_gen;
+mod simple_agg;
 mod sink;
 mod sort;
 mod sort_buffer;
 mod sort_buffer_v0;
 mod sort_v0;
 pub mod source;
+mod stateless_simple_agg;
 mod stream_reader;
 pub mod subtask;
 mod temporal_join;
@@ -117,11 +117,9 @@ pub use dynamic_filter::DynamicFilterExecutor;
 pub use error::{StreamExecutorError, StreamExecutorResult};
 pub use expand::ExpandExecutor;
 pub use filter::FilterExecutor;
-pub use global_simple_agg::SimpleAggExecutor;
 pub use hash_agg::HashAggExecutor;
 pub use hash_join::*;
 pub use hop_window::HopWindowExecutor;
-pub use local_simple_agg::StatelessSimpleAggExecutor;
 pub use lookup::*;
 pub use lookup_union::LookupUnionExecutor;
 pub use merge::MergeExecutor;
@@ -134,9 +132,11 @@ pub use project_set::*;
 pub use rearranged_chain::RearrangedChainExecutor;
 pub use receiver::ReceiverExecutor;
 use risingwave_pb::source::{ConnectorSplit, ConnectorSplits};
+pub use simple_agg::SimpleAggExecutor;
 pub use sink::SinkExecutor;
 pub use sort::*;
 pub use source::*;
+pub use stateless_simple_agg::StatelessSimpleAggExecutor;
 pub use temporal_join::*;
 pub use top_n::{
     AppendOnlyGroupTopNExecutor, AppendOnlyTopNExecutor, GroupTopNExecutor, TopNExecutor,
