+++ src/stream/src/executor/simple_agg.rs
