+++ src/stream/src/executor/stateless_simple_agg.rs
