+++ src/stream/src/from_proto/simple_agg.rs
