+++ src/stream/src/from_proto/stateless_simple_agg.rs
